Modify comparison checker
Assumption is that due to the while loop when fetching the data, the new lottery results will never be empty. Hence there is no need to introduce checking between the old and the new to maintain data integrity.

Comparison has been updated to return a list of updated keys, which will be sent via Firebase. Firebase class has yet to be updated